+++ src/page/Order/index.tsx
@@ -3,15 +3,10 @@ import MessageCardSection from './components/MessageCardSection';
 import MessageInput from './components/MessageInput';
 import SenderInfo from './components/SenderInfo';
 import ReceiverField from './components/ReceiverField';
-import { useNavigate, useParams } from 'react-router-dom';
-import { FormProvider, useForm } from 'react-hook-form';
+import { FormProvider } from 'react-hook-form';
 import toLocaleString from '@/utils/toLocaleString';
-import useRanking from './hooks/useRnaking';
 import ProductInfo from './components/ProductInfo';
-import { useUserInfo } from '@/contexts/UserInfoContext';
-import postOrderInfo from './utils/postOrderInfo';
-import { toast } from 'react-toastify';
-import { ROUTES } from '@/routes/Routes';
+import useOrderForm from './hooks/useOrderForm';
 
 export interface OrderInfoValues {
   message: string;
@@ -20,39 +15,7 @@ export interface OrderInfoValues {
 }
 
 const OrderPage = () => {
-  const { id } = useParams<{ id: string }>();
-  const navigate = useNavigate();
-  const { productSummaryData } = useRanking(id as string);
-  const { userInfo } = useUserInfo();
-  const orderForm = useForm<OrderInfoValues>({
-    defaultValues: { message: '축하해요.', name: userInfo.name, receiverInfos: [] },
-  });
-  const orderData = orderForm.getValues();
-  const price = orderData.receiverInfos.length * (productSummaryData?.price || 0);
-
-  const onSubmit = async () => {
-    if (!id) return;
-    if (orderData.receiverInfos.length === 0) {
-      toast('받는 사람이 없습니다');
-      return;
-    }
-
-    const isSuccess = await postOrderInfo({
-      orderData,
-      navigate,
-      id,
-    });
-
-    if (isSuccess) {
-      alert(`
-      주문이 완료되었습니다.
-      상품명: ${productSummaryData?.name}
-      구매 수량: ${orderData.receiverInfos.length}
-      발신자 이름: ${orderData.name}
-      메시지: ${orderData.message}`);
-      navigate(ROUTES.HOME);
-    }
-  };
+  const { orderForm, onSubmit, price, productSummaryData } = useOrderForm();
 
   return (
     <>
